Fix the issues #66
1. Move `Transformer` class to examples, with the following polish:
  - Remove inner peak of texar.modules
  - Remove the redundant loading of dictionary
  - Remove the magic numbers (bos_token_id, etc)

2. Fix the help information of `run_mode`
3. Remove the `getattr` of beam_width
4. Remove unnecessary `torch.cuda.empty_cache`

Remaining Issues:
1. The usage of texar.data as the substitution of torchtext
2. The CUDA migration of models
