+++ examples/transformer/bleu_tool.py
@@ -118,7 +118,8 @@ def compute_bleu(reference_corpus, translation_corpus,
         ratio = translation_length / reference_length
         if ratio == 0:
             bp = 0
-        bp = math.exp(1 - 1.0 / ratio) if ratio < 1.0 else 1.0
+        else:
+            bp = math.exp(1 - 1.0 / ratio) if ratio < 1.0 else 1.0
     bleu = geo_mean * bp
     return np.float32(bleu)
