+++ examples/transformer/config_wmt14.py
@@ -9,4 +9,4 @@ max_decoding_length = 256
 
 filename_prefix = "processed."
 input_dir = 'temp/run_en_de_bpe/data'
-vocab_file = input_dir + '/processed.vocab.pickle'
+vocab_file = input_dir + '/processed.vocab.text'
